Updated the tests to use the utility files and updated help
