+++ src/Modules/Workspaces/Commands.Workspaces.Test/RemovePowerBIWorkspaceUserTests.cs
@@ -5,10 +5,8 @@
 
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Management.Automation;
-using Microsoft.PowerBI.Api.V2.Models;
-using Microsoft.PowerBI.Commands.Profile;
+using Microsoft.PowerBI.Commands.Profile.Test;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.PowerBI.Commands.Workspaces.Test
@@ -23,16 +21,13 @@ namespace Microsoft.PowerBI.Commands.Workspaces.Test
         {
             using (var ps = System.Management.Automation.PowerShell.Create())
             {
-                ps.AddCommand(new CmdletInfo($"{ConnectPowerBIServiceAccount.CmdletVerb}-{ConnectPowerBIServiceAccount.CmdletName}", typeof(ConnectPowerBIServiceAccount))).AddParameter("Environment", "PPE");
-                var result = ps.Invoke();
-                Assert.IsFalse(ps.HadErrors);
-                Assert.IsNotNull(result);
-                ps.Commands.Clear();
+                ProfileTestUtilities.ConnectToPowerBI(ps);
 
-                var workspace = GetWorkspace(ps);
+                var workspace = WorkspacesTestUtilities.GetWorkspace(ps);
+
                 if (workspace == null)
                 {
-                    return;
+                    Assert.Inconclusive("No workspaces found to perform end to end test");
                 }
 
                 var parameters = new Dictionary<string, object>()
@@ -43,7 +38,7 @@ namespace Microsoft.PowerBI.Commands.Workspaces.Test
                 };
 
                 ps.AddCommand(new CmdletInfo($"{RemovePowerBIWorkspaceUser.CmdletVerb}-{RemovePowerBIWorkspaceUser.CmdletName}", typeof(RemovePowerBIWorkspaceUser))).AddParameters(parameters);
-                result = ps.Invoke();
+                var result = ps.Invoke();
                 Assert.IsNotNull(result);
                 Assert.IsTrue(result.Count > 0);
             }
@@ -56,16 +51,13 @@ namespace Microsoft.PowerBI.Commands.Workspaces.Test
         {
             using (var ps = System.Management.Automation.PowerShell.Create())
             {
-                ps.AddCommand(new CmdletInfo($"{ConnectPowerBIServiceAccount.CmdletVerb}-{ConnectPowerBIServiceAccount.CmdletName}", typeof(ConnectPowerBIServiceAccount))).AddParameter("Environment", "PPE");
-                var result = ps.Invoke();
-                Assert.IsFalse(ps.HadErrors);
-                Assert.IsNotNull(result);
-                ps.Commands.Clear();
+                ProfileTestUtilities.ConnectToPowerBI(ps);
 
-                var group = GetGroup(ps);
+                var group = WorkspacesTestUtilities.GetGroup(ps);
+
                 if (group == null)
                 {
-                    return;
+                    Assert.Inconclusive("No groups found to perform end to end test");
                 }
 
                 var parameters = new Dictionary<string, object>()
@@ -76,7 +68,7 @@ namespace Microsoft.PowerBI.Commands.Workspaces.Test
                 };
 
                 ps.AddCommand(new CmdletInfo($"{RemovePowerBIWorkspaceUser.CmdletVerb}-{RemovePowerBIWorkspaceUser.CmdletName}", typeof(RemovePowerBIWorkspaceUser))).AddParameters(parameters);
-                result = ps.Invoke();
+                var result = ps.Invoke();
                 Assert.IsNotNull(result);
                 Assert.IsTrue(result.Count > 0);
             }
@@ -100,39 +92,6 @@ namespace Microsoft.PowerBI.Commands.Workspaces.Test
                 var result = ps.Invoke();
                 Assert.Fail("Should not have reached this point");
             }
-        }
-
-        //Move this method to a common test utilities file
-        private static Group GetWorkspace(System.Management.Automation.PowerShell ps, string id = null)
-        {
-            ps.AddCommand(new CmdletInfo($"{GetPowerBIWorkspace.CmdletVerb}-{GetPowerBIWorkspace.CmdletName}", typeof(GetPowerBIWorkspace))).AddParameter("Scope", "Organization");
-            var results = ps.Invoke();
-            ps.Commands.Clear();
-
-            if (results.Any())
-            {
-                var workspaces = results.Select(x => (Group)x.BaseObject);
-
-                return id == null ? workspaces.First() : workspaces.First(x => x.Id == id);
-            }
-
-            return null;
-        }
-
-        private static Group GetGroup(System.Management.Automation.PowerShell ps, string id = null)
-        {
-            ps.AddCommand(new CmdletInfo($"{GetPowerBIWorkspace.CmdletVerb}-{GetPowerBIWorkspace.CmdletName}", typeof(GetPowerBIWorkspace))).AddParameter("Scope", "Individual");
-            var results = ps.Invoke();
-            ps.Commands.Clear();
-
-            if (results.Any())
-            {
-                var workspaces = results.Select(x => (Group)x.BaseObject);
-
-                return id == null ? workspaces.First() : workspaces.First(x => x.Id == id);
-            }
-
-            return null;
         }
     }
 }
