+++ src/Modules/Workspaces/Commands.Workspaces/help/Get-PowerBIWorkspace.md
@@ -37,7 +37,6 @@ PS C:\> {{ Add example code here }}
 Type: PowerBIUserScope
 Parameter Sets: (All)
 Aliases:
-Accepted values: Individual, Organization
 
 Required: False
 Position: Named
